Encoder: Add mno-unaligned-access compiler optimization flag
SIGBUS error is observed due to invalid upcasting of a byte pointer
to 32-bit pointer. The fix is to pass “-mno-unaligned-access” pass
to the compiler to prevent the compiler from making some
optimization and be more conservative. This option tells the
compiler to not assume that 32-bit accesses can use instructions
that require alignment.

Change-Id: Ic6c2e61d3b80c9b158eebfef477d0922e4b0ecc9
CRs-Fixed: 1061537